update method delta value now is measured in seconds

+++ project/src/main/java/game/model/entity/Entity.java
@@ -45,5 +45,5 @@ public abstract class Entity implements IEntity {
     public abstract boolean handleCollision(IEntity entity);
 
     @Override
-    public abstract void update(long delta);
+    public abstract void update(double delta);
 }
